+++ src/app/shared/login.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { OidcProvider } from './login/oidcProvider';
-import { environment } from '../environments/environment';
+import { OidcProvider } from '../login/oidcProvider';
+import { environment } from '../../environments/environment';
 
 @Injectable({
   providedIn: 'root',
