+++ src/app/shared/profile.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Profile } from './profile/profile';
+import { Profile } from '../profile/profile';
 import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { environment } from '../environments/environment';
+import { environment } from '../../environments/environment';
 
 @Injectable({
   providedIn: 'root',
